replaced c filenames to match their header counterparts
