+++ inc/powerpc_spu_command_handler.h
@@ -1,0 +1,10 @@
+#ifndef POWERPC_SPU_COMMAND_HANDLER_H
+#define POWERPC_SPU_COMMAND_HANDLER_H
+
+#ifdef RbStatusCodes
+#else
+#include "../inc/amd64_command_ring_buffer.h"
+// have the enum be readily defined for a strategy pattern for enumerated
+// values.
+
+#endif // POWERPC_SPU_COMMAND_HANDLER_H
