+++ src/amd64_comptime_interpreter.c
