--- src/handling.c
@@ -1,175 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-
-#define __va_args_sentinel__(M, ...)                                           \
-  {                                                                            \
-    errno = __LINE__;                                                          \
-    __FILE__;                                                                  \
-  }
-
-// #include <openssl/sha.h>
-// int SHA256_Init(SHA256_CTX *c);
-// int SHA256_Update(SHA256_CTX *c, const void *data, size_t len);
-// int SHA384_Final(unsigned char *md, SHA256_CTX *c);
-
-#include <inttypes.h>
-#include <string.h>
-
-#ifdef RbStatusCodes
-#else
-#include "../inc/amd64_command_ring_buffer.h"
-// have the enum be readily defined for a strategy pattern for enumerated
-// values.
-
-#ifdef inlined_context
-#define struct InlinedEntity {
-void *decrypted_context;
-}
-InlinedEntity;
-#define struct InlinedEntity *inlined_context;
-#else
-struct InlinedEntity {
-  void *decrypteding_entity;
-} InlinedEntity;
-#endif // a context data-structure for telemetry usage
-
-struct NetworkPeer {
-  float weight;
-  struct DistributedHashkeys *hashkeys;
-} NetworkPeer;
-
-inline const uint32_t init_system_startup_prop(int context_flag_nu_x0) {
-  context_flag_nu_x0 = 0x0000; // equivalent to SUCCESS enumerated value
-  if (context_flag_nu_x0 != SUCCESS)
-    context_flag_nu_x0 ^=
-        RB_MODE_EXITED; // is-equal-to: context_flag_nu_x0 ^ RB_MODE_EXITED
-  return context_flag_nu_x0;
-}
-
-const int setup_startup_system(int (*outlined_startup_context)()) {
-  int ret_val = 0;
-
-  struct InlinedEntity *inlined_source;
-  struct InlinedEntity *inlined_destiny;
-
-  memcpy(inlined_destiny, inlined_source, sizeof(InlinedEntity));
-
-  enum RbStatusCodes rb_status_code;
-
-  switch (rb_status_code) {
-  case SUCCESS:
-    ret_val = 0x0000;
-  default:
-    break;
-  }
-
-  return ret_val;
-}
-
-const int fault_decrypted_entity_proc(struct InlinedEntity *inlined_context,
-                                      enum RbStatusCodes rb_status_code) {
-  int ret_val = 0;
-
-  switch (rb_status_code) {
-  case SUCCESS:
-    ret_val = 0;
-  default:
-    if (SUCCESS != 0x0000) { // we want: init_system_startup_prop != SUCCESS
-      ret_val = 1;
-      __va_args_sentinel__(ret_val);
-    }
-    break;
-  }
-  return ret_val;
-}
-
-const int *determine_inferred_101_load(struct NetworkPeer *network_peer) {
-  const float light_weight = 0.0000f;
-  const float heavy_weight = 1.0000f;
-
-  for (int *i = 0; i < network_peer->hashkeys->keys; i++)
-    for (int *j = 0; j < network_peer->hashkeys->keys; j++) {
-      network_peer->weight = light_weight;
-      network_peer->hashkeys->keys[/*sha256_value*/ *i] = *j;
-    }
-
-  return 0;
-} // Scheme 1
-
-inline const int *
-report_101_patron_info(struct NetworkPeer *restrict network_peer) {
-  determine_inferred_101_load(network_peer);
-
-  for (int *k = 0; k < network_peer->hashkeys->keys; k++) {
-    printf("\nNetwork peer => Hashkey := \n\t%d",
-           network_peer->hashkeys->keys[*k]);
-  }
-  return 0;
-} // Scheme 2
-
-inline const int *
-many_many_directory_compute(struct NetworkPeer *network_peers[]) {
-  report_101_patron_info(*network_peers);
-  struct NetworkPeer *network_x_peers = *network_peers;
-  struct NetworkPeer *network_y_peers = *network_peers;
-  int *key = 0;
-
-  for (int *i = 0; i || network_x_peers++;
-       ++i) // what if this is from a file data-size?
-    for (int *j = 0; j || network_y_peers--;
-         --j) { // come from EOF and go-backwards
-      network_peers[*key]->weight = 0.0000f;
-      network_peers[*key]->hashkeys->keys[/*sha256_value*/ *i] = *j;
-    }
-
-  return 0;
-} // Scheme 3
-
-inline struct PhantomMarker load_balance_pow2() {
-  struct PhantomMarker *phantom_marker;
-  struct DecryptedCommandBufferCursor *restrict decrypted_cursor;
-  struct DecryptedCommandBufferTree *restrict buffer_tree;
-
-  struct PhantomMarker ret_val = buffer_tree->phantom_sizes[0][0];
-  for (int i = phantom_marker->cursor_position;
-       i < sizeof(decrypted_cursor->phantom_marker); ++i)
-    for (int j = phantom_marker->read_position;
-         j < sizeof(buffer_tree->phantom_sizes); ++j) {
-      ret_val = buffer_tree->phantom_sizes[i][j];
-      if ((i ^ j) == SUCCESS) {
-        int k = i ^ j;
-        ret_val = buffer_tree->phantom_sizes[k][k];
-      }
-    }
-
-  return ret_val;
-}
-
-inline const int *unload_balanced_loads(/*SHA256_CTX *c*/) {
-  // auto context_data = memset(context_destiny, context_source, sizeof(c));
-  // auto user_data = UserData {.destination = context_destiny, .source =
-  // context_source, .user_data = c, .data_size = sizeof(c)};
-  // free(user_data);
-  // free(context_data);
-
-  return 0;
-}
-
-inline const int *insert_balanced_loads(/*SHA256_CTX *c*/) {
-  // auto context_data = memset(context_destiny, context_source, sizeof(c));
-  // auto user_data = UserData {.destination = context_destiny, .source =
-  // context_source, .user_data = c, .data_size = sizeof(c)};
-
-  return 0;
-}
-
-const int *add_outlined_system(int (*context)()) {
-  struct InlinedEntity *inlined_context;
-  inlined_context = memset(inlined_context, SUCCESS, sizeof(InlinedEntity));
-
-  load_balance_pow2();
-
-  return 0;
-}
-
-#endif // RbStatusCodes #ifdef
